pass menuitems and theme context as props
