+++ src/components/Menu/index.jsx
@@ -1,20 +1,8 @@
-import { useContext } from "react";
 import MenuItem from "../MenuItem";
 import ThemeSwitcher from "../ThemeSwitcher";
 import styles from "./Menu.module.css";
-import { ThemeContext } from "../../context/ThemeContext";
 
-const Menu = ({ homeRef, aboutRef, contactRef, skillsRef, projectsRef }) => {
-  const { darkMode, switchTheme } = useContext(ThemeContext);
-
-  const menuItems = [
-    { title: "Home", reference: homeRef },
-    { title: "About", reference: aboutRef },
-    { title: "Skills", reference: skillsRef },
-    { title: "Projects", reference: projectsRef },
-    { title: "Contact", reference: contactRef },
-  ];
-
+const Menu = ({ menuItems, darkMode, switchTheme }) => {
   return (
     <div className={styles.menu}>
       {menuItems.map((item) => (
